Abstract file writing into Utils
