+++ frontend/src/Language/Granule/Syntax/Preprocessor.hs
@@ -1,18 +1,13 @@
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Language.Granule.Syntax.Preprocessor where
 
-import Control.Exception (SomeException, throwIO, try)
-import Control.Monad (when)
 import Data.List (intercalate)
-import System.Directory (removeFile, renameFile)
-import System.FilePath (splitFileName)
-import System.IO (hClose, hPutStr, openTempFile)
 
 import Language.Granule.Syntax.Preprocessor.Latex
 import Language.Granule.Syntax.Preprocessor.Markdown
 
+import Language.Granule.Utils
 
 -- | Preprocess the source file based on the file extension.
 preprocess :: Maybe (String -> String) -> Bool -> String -> FilePath -> IO String
@@ -23,21 +18,8 @@ preprocess mbRewriter keepOldFile file env
         case mbRewriter of
           Just rewriter -> do
             let processedSrc = preprocessOnlyGranule rewriter src
-            -- open a temporary file
-            (tempFile, tempHd) <- uncurry openTempFile (splitFileName file)
-            -- write the processed source to the temporary file
-            try (hPutStr tempHd processedSrc) >>= \case
-              Right () -> do
-                hClose tempHd
-                -- if we are keeping the original source file, then rename it
-                when keepOldFile (renameFile file (file <> ".bak"))
-                -- move the temp file to the original source file path
-                renameFile tempFile file
-                return $ stripNonGranule processedSrc
-              Left (e :: SomeException) -> do
-                hClose tempHd
-                removeFile tempFile
-                throwIO e
+            written <- writeSrcFile file keepOldFile processedSrc
+            return $ stripNonGranule written
           Nothing -> return $ stripNonGranule src
       Nothing -> error
         $ "Unrecognised file extension: "
